# doc/api.yaml
{"openapi":"3.0.0","paths":{"/user":{"post":{"operationId":"UsersController_create","summary":"","description":"Create user","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"Create user","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"400":{"description":"Bad request. body does not contain required fields"}},"tags":["User"]},"get":{"operationId":"UsersController_findAll","summary":"","description":"Get all users","parameters":[],"responses":{"200":{"description":"Get all users","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}}},"tags":["User"]}},"/user/{id}":{"get":{"operationId":"UsersController_findOne","summary":"","description":"Get user by id","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Get user by id","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"400":{"description":"Bad request. userId is invalid (not uuid)"},"404":{"description":"User is not found"}},"tags":["User"]},"put":{"operationId":"UsersController_update","summary":"","description":"Updates a user password by id","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":"Updates a user password by id","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"400":{"description":"Bad request. userId is invalid (not uuid)"},"403":{"description":"Wrong old password"},"404":{"description":"User not found"}},"tags":["User"]},"delete":{"operationId":"UsersController_remove","summary":"","description":"Delete user","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":"Remove user"},"400":{"description":"Bad request. userId is invalid (not uuid)"},"404":{"description":"User not found"}},"tags":["User"]}},"/artist":{"post":{"operationId":"ArtistController_create","summary":"","description":"Create artist","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateArtistDto"}}}},"responses":{"201":{"description":"Create artist","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Artist"}}}},"400":{"description":"Bad request. body does not contain required fields"}},"tags":["Artist"]},"get":{"operationId":"ArtistController_findAll","summary":"","description":"Get all artists","parameters":[],"responses":{"200":{"description":"Get all artists","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Artist"}}}}}},"tags":["Artist"]}},"/artist/{id}":{"get":{"operationId":"ArtistController_findOne","summary":"","description":"Get artist by id","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Get artist by id","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Artist"}}}},"400":{"description":"Bad request. artistId is invalid (not uuid)"},"404":{"description":"Artist is not found"}},"tags":["Artist"]},"put":{"operationId":"ArtistController_update","summary":"","description":"Updates a artist by id","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateArtistDto"}}}},"responses":{"200":{"description":"Updates a artist by id","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Artist"}}}},"400":{"description":"Bad request. artistId is invalid (not uuid)"},"404":{"description":"Artist is not found"}},"tags":["Artist"]},"delete":{"operationId":"ArtistController_remove","summary":"","description":"Updates a artist by id","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":"Remove artist"},"400":{"description":"Bad request. artistId is invalid (not uuid)"},"404":{"description":"Artist is not found"}},"tags":["Artist"]}},"/favs":{"get":{"operationId":"FavoritesController_findAll","summary":"","description":"Get all favorites","parameters":[],"responses":{"200":{"description":"Get all favorites","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Favorite"}}}}},"tags":["Favorites"]}},"/favs/track/{id}":{"post":{"operationId":"FavoritesController_addTrack","summary":"","description":"Add track to favorites","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":"Add track to favorites"},"400":{"description":"Bad request. trackId is invalid (not uuid)"},"422":{"description":"Track is not found"}},"tags":["Favorites"]},"delete":{"operationId":"FavoritesController_removeTrack","summary":"","description":"Remove track from favorites","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":"Remove track from favorites"},"400":{"description":"Bad request. trackId is invalid (not uuid)"},"404":{"description":"Track is not found"}},"tags":["Favorites"]}},"/favs/album/{id}":{"post":{"operationId":"FavoritesController_addAlbum","summary":"","description":"Add album to favorites","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":"Add album to favorites"},"400":{"description":"Bad request. albumId is invalid (not uuid)"},"422":{"description":"Album is not found"}},"tags":["Favorites"]},"delete":{"operationId":"FavoritesController_removeAlbum","summary":"","description":"Remove album from favorites","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":"Remove album from favorites"},"400":{"description":"Bad request. albumId is invalid (not uuid)"},"404":{"description":"Album is not found"}},"tags":["Favorites"]}},"/favs/artist/{id}":{"post":{"operationId":"FavoritesController_addArtist","summary":"","description":"Add artist to favorites","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":"Add artist to favorites"},"400":{"description":"Bad request. artistId is invalid (not uuid)"},"422":{"description":"Artist is not found"}},"tags":["Favorites"]},"delete":{"operationId":"FavoritesController_removeArtist","summary":"","description":"Remove artist from favorites","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":"Remove artist from favorites"},"400":{"description":"Bad request. artistId is invalid (not uuid)"},"404":{"description":"Artist is not found"}},"tags":["Favorites"]}},"/track":{"post":{"operationId":"TrackController_create","summary":"","description":"Create track","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTrackDto"}}}},"responses":{"201":{"description":"Create track","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Track"}}}},"400":{"description":"Bad request. body does not contain required fields"}},"tags":["Track"]},"get":{"operationId":"TrackController_findAll","summary":"","description":"Get all tracks","parameters":[],"responses":{"200":{"description":"Get all tracks","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Track"}}}}}},"tags":["Track"]}},"/track/{id}":{"get":{"operationId":"TrackController_findOne","summary":"","description":"Get track by id","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Get track by id","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Track"}}}},"400":{"description":"Bad request. trackId is invalid (not uuid)"},"404":{"description":"Track is not found"}},"tags":["Track"]},"put":{"operationId":"TrackController_update","summary":"","description":"Updates a track by id","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTrackDto"}}}},"responses":{"200":{"description":"Updates a track by id","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Track"}}}},"400":{"description":"Bad request. trackId is invalid (not uuid)"},"404":{"description":"Track is not found"}},"tags":["Track"]},"delete":{"operationId":"TrackController_remove","summary":"","description":"Updates a track by id","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":"Remove track"},"400":{"description":"Bad request. trackId is invalid (not uuid)"},"404":{"description":"Track is not found"}},"tags":["Track"]}},"/album":{"post":{"operationId":"AlbumController_create","summary":"","description":"Create album","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAlbumDto"}}}},"responses":{"201":{"description":"Create album","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Album"}}}},"400":{"description":"Bad request. body does not contain required fields"}},"tags":["Album"]},"get":{"operationId":"AlbumController_findAll","summary":"","description":"Get all albums","parameters":[],"responses":{"200":{"description":"Get all albums","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Album"}}}}}},"tags":["Album"]}},"/album/{id}":{"get":{"operationId":"AlbumController_findOne","summary":"","description":"Get album by id","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Get album by id","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Album"}}}},"400":{"description":"Bad request. albumId is invalid (not uuid)"},"404":{"description":"Album is not found"}},"tags":["Album"]},"put":{"operationId":"AlbumController_update","summary":"","description":"Updates a album by id","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateAlbumDto"}}}},"responses":{"200":{"description":"Updates a album by id","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Track"}}}},"400":{"description":"Bad request. albumId is invalid (not uuid)"},"404":{"description":"Album is not found"}},"tags":["Album"]},"delete":{"operationId":"AlbumController_remove","summary":"","description":"Updates a album by id","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":"Remove album"},"400":{"description":"Bad request. albumId is invalid (not uuid)"},"404":{"description":"Album is not found"}},"tags":["Album"]}}},"info":{"title":"Basic service","description":"Basic api","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"CreateUserDto":{"type":"object","properties":{"login":{"type":"string","example":"NewUser"},"password":{"type":"string","example":"password"}},"required":["login","password"]},"User":{"type":"object","properties":{"id":{"type":"string","example":"3fa85f64-5717-4562-b3fc-2c963f66afa6"},"login":{"type":"string","example":"New user"},"version":{"type":"number","example":1},"createdAt":{"format":"date-time","type":"string","example":1655000000},"updatedAt":{"format":"date-time","type":"string","example":1655000000}},"required":["id","login","version","createdAt","updatedAt"]},"UpdateUserDto":{"type":"object","properties":{"oldPassword":{"type":"string","example":"password"},"newPassword":{"type":"string","example":"new_password"}},"required":["oldPassword","newPassword"]},"CreateArtistDto":{"type":"object","properties":{"name":{"type":"string","example":"Artist 1"},"grammy":{"type":"boolean","example":true}},"required":["name"]},"Artist":{"type":"object","properties":{"id":{"type":"string","example":"3fa85f64-5717-4562-b3fc-2c963f66afa6"},"name":{"type":"string","example":"Artist 1"},"grammy":{"type":"boolean","example":true}},"required":["id","name"]},"UpdateArtistDto":{"type":"object","properties":{"name":{"type":"string","example":"Artist 1"},"grammy":{"type":"boolean","example":true}},"required":["name"]},"Favorite":{"type":"object","properties":{"tracks":{"example":[{"id":"3fa85f64-5717-4562-b3fc-2c963f66afa6","name":"Track 1","duration":10}],"type":"array","items":{"type":"string"}},"albums":{"example":[{"id":"3fa85f64-5717-4562-b3fc-2c963f66afa6","name":"Album 1","year":1970,"artistId":"3fa85f64-5717-4562-b3fc-2c963f66afa6"}],"type":"array","items":{"type":"string"}},"artists":{"example":[{"id":"3fa85f64-5717-4562-b3fc-2c963f66afa6","name":"Artist 1","grammy":true}],"type":"array","items":{"type":"string"}}},"required":["tracks","albums","artists"]},"CreateTrackDto":{"type":"object","properties":{"name":{"type":"string","example":"Track 1"},"artistId":{"type":"string","example":"3fa85f64-5717-4562-b3fc-2c963f66afa6"},"albumId":{"type":"string","example":"3fa85f64-5717-4562-b3fc-2c963f66afa6"},"duration":{"type":"number","example":10}},"required":["name","duration"]},"Track":{"type":"object","properties":{"id":{"type":"string","example":"3fa85f64-5717-4562-b3fc-2c963f66afa6"},"name":{"type":"string","example":"Track 1"},"artistId":{"type":"string","example":"3fa85f64-5717-4562-b3fc-2c963f66afa6"},"albumId":{"type":"string","example":"3fa85f64-5717-4562-b3fc-2c963f66afa6"},"duration":{"type":"number","example":10}},"required":["id","name","artistId","albumId","duration"]},"UpdateTrackDto":{"type":"object","properties":{"name":{"type":"string","example":"Track 1"},"artistId":{"type":"string","example":"3fa85f64-5717-4562-b3fc-2c963f66afa6"},"albumId":{"type":"string","example":"3fa85f64-5717-4562-b3fc-2c963f66afa6"},"duration":{"type":"number","example":10}},"required":["name","duration"]},"CreateAlbumDto":{"type":"object","properties":{"name":{"type":"string","example":"Album 1"},"year":{"type":"number","example":1970},"artistId":{"type":"string","example":"3fa85f64-5717-4562-b3fc-2c963f66afa6"}},"required":["name","year","artistId"]},"Album":{"type":"object","properties":{"id":{"type":"string","example":"3fa85f64-5717-4562-b3fc-2c963f66afa6"},"name":{"type":"string","example":"Album 1"},"year":{"type":"number","example":1970},"artistId":{"type":"string","example":"3fa85f64-5717-4562-b3fc-2c963f66afa6"}},"required":["id","name","year","artistId"]},"UpdateAlbumDto":{"type":"object","properties":{"name":{"type":"string","example":"Album 1"},"year":{"type":"number","example":1970},"artistId":{"type":"string","example":"3fa85f64-5717-4562-b3fc-2c963f66afa6"}},"required":["name","year","artistId"]}}}}