# src/modules/track/track.controller.ts
import {
  Body,
  Controller,
  Delete,
  Get,
  HttpCode,
  Param,
  Post,
  Put,
} from '@nestjs/common';
import { TrackService } from './track.service';
import { CreateTrackDto } from './dto/create-track.dto';
import { UpdateTrackDto } from './dto/update-track.dto';
import { ApiOperation, ApiResponse, ApiTags } from '@nestjs/swagger';
import { Track } from './entities/track.entity';

@Controller('track')
@ApiTags('Track')
export class TrackController {
  constructor(private readonly trackService: TrackService) {}

  @Post()
  @ApiOperation({ description: 'Create track' })
  @ApiResponse({
    status: 201,
    description: 'Create track',
    type: Track,
  })
  @ApiResponse({
    status: 400,
    description: 'Bad request. body does not contain required fields',
  })
  async create(@Body() createTrackDto: CreateTrackDto) {
    return await this.trackService.create(createTrackDto);
  }

  @Get()
  @ApiOperation({ description: 'Get all tracks' })
  @ApiResponse({
    status: 200,
    description: 'Get all tracks',
    type: Track,
    isArray: true,
  })
  async findAll() {
    return await this.trackService.findAll();
  }

  @Get(':id')
  @ApiOperation({ description: 'Get track by id' })
  @ApiResponse({
    status: 200,
    description: 'Get track by id',
    type: Track,
  })
  @ApiResponse({
    status: 400,
    description: 'Bad request. trackId is invalid (not uuid)',
  })
  @ApiResponse({
    status: 404,
    description: 'Track is not found',
  })
  async findOne(@Param('id') id: string) {
    return await this.trackService.findOne(id);
  }

  @Put(':id')
  @ApiOperation({ description: 'Updates a track by id' })
  @ApiResponse({
    status: 200,
    description: 'Updates a track by id',
    type: Track,
  })
  @ApiResponse({
    status: 400,
    description: 'Bad request. trackId is invalid (not uuid)',
  })
  @ApiResponse({
    status: 404,
    description: 'Track is not found',
  })
  async update(
    @Param('id') id: string,
    @Body() updateTrackDto: UpdateTrackDto,
  ) {
    return await this.trackService.update(id, updateTrackDto);
  }

  @Delete(':id')
  @HttpCode(204)
  @ApiOperation({ description: 'Updates a track by id' })
  @ApiResponse({
    status: 204,
    description: 'Remove track',
  })
  @ApiResponse({
    status: 400,
    description: 'Bad request. trackId is invalid (not uuid)',
  })
  @ApiResponse({
    status: 404,
    description: 'Track is not found',
  })
  async remove(@Param('id') id: string) {
    return await this.trackService.remove(id);
  }
}